+++ android/App/src/main/kotlin/com/darakeon/dfm/base/UIHandler.kt
@@ -2,19 +2,7 @@ package com.darakeon.dfm.base
 
 import android.app.Activity
 import android.app.Dialog
-import android.content.Context
-import android.content.pm.ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE
-import android.content.pm.ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
-import android.content.pm.ActivityInfo.SCREEN_ORIENTATION_REVERSE_LANDSCAPE
-import android.content.pm.ActivityInfo.SCREEN_ORIENTATION_REVERSE_PORTRAIT
 import android.content.pm.ActivityInfo.SCREEN_ORIENTATION_UNSPECIFIED
-import android.content.res.Configuration.ORIENTATION_LANDSCAPE
-import android.content.res.Configuration.ORIENTATION_PORTRAIT
-import android.os.Build
-import android.view.Surface.ROTATION_0
-import android.view.Surface.ROTATION_180
-import android.view.Surface.ROTATION_270
-import android.view.Surface.ROTATION_90
 import android.view.WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON
 import com.darakeon.dfm.dialogs.createWaitDialog
 
@@ -45,38 +33,9 @@ open class UIHandler(
 	}
 
 	private fun disableRotation() {
-		val orientation = activity.resources.configuration.orientation
-		val rotation = display(activity)?.rotation
-
 		activity.requestedOrientation =
-			rotations[rotation]?.get(orientation)
-				?: SCREEN_ORIENTATION_UNSPECIFIED
+			activity.resources.configuration.orientation
 	}
-
-	private fun display(context: Context) =
-		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-			context.display
-		} else {
-			@Suppress("DEPRECATION")
-			activity.windowManager.defaultDisplay
-		}
-
-	private val upSide = hashMapOf(
-		ORIENTATION_PORTRAIT to SCREEN_ORIENTATION_PORTRAIT,
-		ORIENTATION_LANDSCAPE to SCREEN_ORIENTATION_LANDSCAPE
-	)
-
-	private val downSide = hashMapOf(
-		ORIENTATION_PORTRAIT to SCREEN_ORIENTATION_REVERSE_PORTRAIT,
-		ORIENTATION_LANDSCAPE to SCREEN_ORIENTATION_REVERSE_LANDSCAPE
-	)
-
-	private val rotations = hashMapOf(
-		ROTATION_0 to upSide,
-		ROTATION_90 to downSide,
-		ROTATION_180 to downSide,
-		ROTATION_270 to upSide
-	)
 
 	open fun endUIWait() {
 		ended++